Move RequestDispatcher interface and Response to a separate module
This avoids circular dependency issues:
 * helios-client needs RequestDispatcher
 * helios-client depends on helios-authentication
 * helios-authentication needs RequestDispatcher
