+++ helios-client/pom.xml
@@ -17,6 +17,11 @@
 
   <dependencies>
     <dependency>
+      <groupId>com.spotify</groupId>
+      <artifactId>helios-transport</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
       <groupId>com.spotify.docgenerator</groupId>
       <artifactId>scanner</artifactId>
       <version>0.0.1</version>
