+++ pom.xml
@@ -121,6 +121,7 @@
     <module>helios-tools</module>
     <module>helios-testing</module>
     <module>helios-testing-common</module>
+    <module>helios-transport</module>
     <module>helios-service-registration</module>
     <module>helios-services</module>
     <module>helios-system-tests</module>
